+++ crates/proto/src/udp/udp_client_stream.rs
@@ -343,7 +343,7 @@ impl SingleUseUdpSocket {
 }
 
 #[cfg(test)]
-#[cfg(feature = "tokio-compat")]
+#[cfg(feature = "tokio-runtime")]
 mod tests {
     #![allow(clippy::dbg_macro, clippy::print_stdout)]